+++ src/prettier.ts
@@ -1,3 +1,7 @@
+import { throwIfVersionNotInRange } from "./validate-version.js";
+
+throwIfVersionNotInRange("prettier", 2);
+
 export const prettier = {
   printWidth: 120,
 };